Reorganize batch test structure
