+++ tests/conftest.py
@@ -1,0 +1,13 @@
+import pytest
+
+from rest_framework import test
+
+
+@pytest.fixture
+def api_rf() -> test.APIRequestFactory:
+    return test.APIRequestFactory()
+
+
+@pytest.fixture
+def api_client() -> test.APIClient:
+    return test.APIClient()
